Add the missing Copyright and license headers in the libcrypto/comp directory. Requested some time ago by tb@.
According to OpenSSL git history, the original version of this code
appeared in SSLeay 0.9.1b (July 6, 1998).  The LICENSE file in that
release states that the Copyright of SSLeay belongs to Eric Young,
and we believe that Eric still maintained SSLeay himself at that time.
We have seen a small number of examples where Eric credited outside
contributors for code that he included in his distribution, including
citing Copyright notices and license headers as appropriate.  We
found no such hints regarding this code, so it is reasonable to
assume that he wrote this code himself.

Regarding subsequent changes and additions, i inspected the OpenSSL
git repository.

No code change; only Copyright and license comments are added.